extract LinkControllerCoordinator; DI
