+++ paymentsheet/src/main/java/com/stripe/android/link/LinkController.kt
@@ -4,81 +4,36 @@ import android.app.Activity
 import android.content.Context
 import android.os.Parcelable
 import androidx.activity.ComponentActivity
-import androidx.activity.result.ActivityResultLauncher
-import androidx.activity.result.ActivityResultRegistryOwner
 import androidx.annotation.DrawableRes
 import androidx.annotation.RestrictTo
-import androidx.lifecycle.DefaultLifecycleObserver
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.lifecycleScope
-import androidx.lifecycle.repeatOnLifecycle
 import com.stripe.android.common.configuration.ConfigurationDefaults
+import com.stripe.android.link.injection.LinkControllerScope
 import com.stripe.android.model.PaymentMethod
 import com.stripe.android.paymentsheet.PaymentSheet
 import dev.drewhamilton.poko.Poko
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.launch
 import kotlinx.parcelize.Parcelize
 import javax.inject.Inject
 
+@LinkControllerScope
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 class LinkController @Inject internal constructor(
     activity: Activity,
-    private val lifecycleOwner: LifecycleOwner,
-    activityResultRegistryOwner: ActivityResultRegistryOwner,
+    private val linkControllerCoordinator: LinkControllerCoordinator,
     private val viewModel: LinkControllerViewModel,
-    private val selectedPaymentMethodCallback: PresentPaymentMethodsCallback,
-    private val lookupConsumerCallback: LookupConsumerCallback,
-    private val createPaymentMethodCallback: CreatePaymentMethodCallback,
-    linkActivityContract: NativeLinkActivityContract,
 ) {
-
     val state: StateFlow<State> = viewModel.state(activity)
-
-    private val linkActivityResultLauncher: ActivityResultLauncher<LinkActivityContract.Args> =
-        activityResultRegistryOwner.activityResultRegistry.register(
-            key = "LinkController_LinkActivityResultLauncher",
-            contract = linkActivityContract,
-        ) { result ->
-            viewModel.onPresentPaymentMethodsActivityResult(result)
-        }
-
-    init {
-        lifecycleOwner.lifecycleScope.launch {
-            lifecycleOwner.lifecycle.repeatOnLifecycle(Lifecycle.State.STARTED) {
-                launch {
-                    viewModel.presentPaymentMethodsResultFlow
-                        .collect(selectedPaymentMethodCallback::onPresentPaymentMethodsResult)
-                }
-                launch {
-                    viewModel.lookupConsumerResultFlow
-                        .collect(lookupConsumerCallback::onLookupConsumerResult)
-                }
-
-                launch {
-                    viewModel.createPaymentMethodResultFlow
-                        .collect(createPaymentMethodCallback::onCreatePaymentMethodResult)
-                }
-            }
-        }
-
-        lifecycleOwner.lifecycle.addObserver(
-            object : DefaultLifecycleObserver {
-                override fun onDestroy(owner: LifecycleOwner) {
-                    linkActivityResultLauncher.unregister()
-                }
-            }
-        )
-    }
 
     fun configure(configuration: Configuration) {
         viewModel.configure(configuration)
     }
 
     fun presentPaymentMethods(email: String?) {
-        viewModel.onPresent(linkActivityResultLauncher, email)
+        viewModel.onPresentPaymentMethods(
+            launcher = linkControllerCoordinator.linkActivityResultLauncher,
+            email = email
+        )
     }
 
     fun createPaymentMethod() {
