+++ paymentsheet/src/main/java/com/stripe/android/link/injection/LinkControllerScope.kt
@@ -1,0 +1,7 @@
+package com.stripe.android.link.injection
+
+import javax.inject.Scope
+
+@Scope
+@Retention(AnnotationRetention.RUNTIME)
+internal annotation class LinkControllerScope
